fix: organized code + bug fixes
